split source file -> operation.go/definition_helper_funcs.go
